make jbob a library crate
